Conflicts:
	src/gc/fixed_allocator.c
	src/gc/gc_ms2.c

+++ src/gc/system.c
@@ -338,7 +338,7 @@ get_max_pmc_address(ARGIN_NULLOK(const Memory_Pools *mem_pools))
     ASSERT_ARGS(get_max_pmc_address)
     return mem_pools
             ? mem_pools->pmc_pool->end_arena_memory
-            : -1;
+            : (size_t)-1;
 }
 
 
@@ -556,5 +556,5 @@ F<src/gc/api.c> and F<include/parrot/gc_api.h>.
  * Local variables:
  *   c-file-style: "parrot"
  * End:
- * vim: expandtab shiftwidth=4:
+ * vim: expandtab shiftwidth=4 cinoptions='\:2=2' :
  */
